+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Click up arrow to go back to TOC.
 
 - Add extra environment variables check when finding dependencies. [issue #91](https://github.com/BingLingGroup/autosub/issues/91)
 - Add arguments parser input when open it without arguments. [issue #92](https://github.com/BingLingGroup/autosub/issues/92)
+- Add subtitles processing when not input `-SRC`.
 
 #### Changed(Unreleased)